requirements.txt
playwright
python-dotenv
requests
openai
httpx
